+++ srcs/raycasting/raycasting.c
@@ -6,7 +6,7 @@
 /*   By: myokono <myokono@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/23 10:40:24 by myokono           #+#    #+#             */
-/*   Updated: 2025/04/30 18:46:45 by myokono          ###   ########.fr       */
+/*   Updated: 2025/04/30 18:52:03 by myokono          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -32,41 +32,6 @@ static void	perform_dda(t_game *game)
 			&&game->ray.map_x >= 0 && game->ray.map_x < game->map.width
 			&&game->map.grid[game->ray.map_y][game->ray.map_x] == '1')
 			game->ray.hit = 1;
-	}
-}
-
-void	calculate_wall_distance(t_game *game)
-{
-	if (game->ray.side == 0)
-		game->ray.perp_wall_dist = (game->ray.map_x - game->player.pos_x
-				+ (1 - game->ray.step_x) / 2) / game->ray.ray_dir_x;
-	else
-		game->ray.perp_wall_dist = (game->ray.map_y - game->player.pos_y
-				+ (1 - game->ray.step_y) / 2) / game->ray.ray_dir_y;
-}
-
-void	calculate_wall_height(t_game *game)
-{
-	game->ray.line_height = (int)(WINDOW_HEIGHT / game->ray.perp_wall_dist);
-	game->ray.draw_start = -game->ray.line_height / 2 + WINDOW_HEIGHT / 2;
-	if (game->ray.draw_start < 0)
-		game->ray.draw_start = 0;
-	game->ray.draw_end = game->ray.line_height / 2 + WINDOW_HEIGHT / 2;
-	if (game->ray.draw_end >= WINDOW_HEIGHT)
-		game->ray.draw_end = WINDOW_HEIGHT - 1;
-	if (game->ray.side == 0)
-	{
-		if (game->ray.ray_dir_x > 0)
-			game->ray.tex_num = EAST;
-		else
-			game->ray.tex_num = WEST;
-	}
-	else
-	{
-		if (game->ray.ray_dir_y > 0)
-			game->ray.tex_num = SOUTH;
-		else
-			game->ray.tex_num = NORTH;
 	}
 }
